test(tests): address review feedback for precompile coinbase
Move the test into test_precompiles.py, tighten the docstring to
state-transition only, and drop the hardcoded gas_limit.
